requirements.txt
opencv-python
mediapipe
torch
torchvision
albumentations
tqdm
pandas
scikit-learn
matplotlib
pyyaml
onnx
onnxruntime
